Homework-10 Основы Hibernate - как аннотации-hibernate влияют на формирование sql-запросов - Классы Client, Address, Phone размечены таким образом, что при сохранении/чтении объекта Client каскадно сохраняются/читаются вложенные объекты; - Hibernate создаёт только 3 таблицы: для телефонов, адресов и клиентов; - При сохранении нового объекта запросы на update отсутствуют; - Все тесты проходят.
