--- L21-jpql/homework-template/build.gradle.kts
@@ -1,19 +1,0 @@
-dependencies {
-    implementation ("org.projectlombok:lombok")
-    annotationProcessor ("org.projectlombok:lombok")
-
-    implementation("ch.qos.logback:logback-classic")
-    implementation("org.hibernate.orm:hibernate-core")
-    implementation("org.flywaydb:flyway-core")
-
-    implementation("org.postgresql:postgresql")
-
-    testImplementation("com.h2database:h2")
-    testImplementation("org.junit.jupiter:junit-jupiter-engine")
-    testImplementation("org.junit.jupiter:junit-jupiter-params")
-    testImplementation("org.assertj:assertj-core")
-    testImplementation("org.mockito:mockito-junit-jupiter")
-
-    testImplementation("org.testcontainers:junit-jupiter")
-    testImplementation("org.testcontainers:postgresql")
-}
