--- L21-jpql/homework-template/src/main/java/ru/otus/crm/model/Address.java
@@ -1,7 +1,0 @@
-package ru.otus.crm.model;
-
-public class Address {
-    public Address(Long id, String street) {
-        throw new UnsupportedOperationException();
-    }
-}
