--- L21-jpql/homework-template/src/main/java/ru/otus/crm/model/Phone.java
@@ -1,7 +1,0 @@
-package ru.otus.crm.model;
-
-public class Phone {
-    public Phone(Long id, String number) {
-        throw new UnsupportedOperationException();
-    }
-}
