--- L21-jpql/homework-template/src/main/resources/db/migration/V1__initial_schema.sql
@@ -1,18 +1,0 @@
--- Для @GeneratedValue(strategy = GenerationType.IDENTITY)
-/*
-create table client
-(
-    id   bigserial not null primary key,
-    name varchar(50)
-);
-
- */
-
--- Для @GeneratedValue(strategy = GenerationType.SEQUENCE)
-create sequence client_SEQ start with 1 increment by 1;
-
-create table client
-(
-    id   bigint not null primary key,
-    name varchar(50)
-);
