--- L21-jpql/homework-template/src/test/java/ru/otus/core/repository/DataTemplateHibernateTest.java
@@ -1,81 +1,0 @@
-package ru.otus.core.repository;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Optional;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import ru.otus.base.AbstractHibernateTest;
-import ru.otus.crm.model.Client;
-
-@SuppressWarnings("java:S125")
-class DataTemplateHibernateTest extends AbstractHibernateTest {
-
-    @Test
-    @Disabled("Удалить при выполнении ДЗ")
-    @DisplayName(" корректно сохраняет, изменяет и загружает клиента по заданному id")
-    void shouldSaveAndFindCorrectClientById() {
-        // given
-        var client = new Client("Вася");
-
-        // Это надо раскомментировать, у выполненного ДЗ, все тесты должны проходить
-        // Кроме удаления комментирования, тестовый класс менять нельзя
-        /*
-                var client = new Client(null, "Vasya", new Address(null, "AnyStreet"), List.of(new Phone(null, "13-555-22"),
-                        new Phone(null, "14-666-333")));
-        */
-
-        // when
-        var savedClient = transactionManager.doInTransaction(session -> {
-            clientTemplate.insert(session, client);
-            return client;
-        });
-
-        // then
-        assertThat(savedClient.getId()).isNotNull();
-        assertThat(savedClient.getName()).isEqualTo(client.getName());
-
-        // when
-        var loadedSavedClient = transactionManager.doInReadOnlyTransaction(
-                session -> clientTemplate.findById(session, savedClient.getId()).map(Client::clone));
-
-        // then
-        assertThat(loadedSavedClient)
-                .isPresent()
-                .get()
-                .usingRecursiveComparison()
-                .isEqualTo(savedClient);
-
-        // when
-        savedClient.setName("updatedName");
-        transactionManager.doInTransaction(session -> {
-            clientTemplate.update(session, savedClient);
-            return null;
-        });
-
-        // then
-        Optional<Client> loadedClient = transactionManager.doInReadOnlyTransaction(
-                session -> clientTemplate.findById(session, savedClient.getId()).map(Client::clone));
-        assertThat(loadedClient).isPresent();
-        assertThat(loadedClient).get().usingRecursiveComparison().isEqualTo(savedClient);
-
-        // when
-        var clientList = transactionManager.doInReadOnlyTransaction(session ->
-                clientTemplate.findAll(session).stream().map(Client::clone).toList());
-
-        // then
-        assertThat(clientList).hasSize(1);
-        assertThat(clientList.getFirst()).usingRecursiveComparison().isEqualTo(savedClient);
-
-        // when
-        clientList = transactionManager.doInReadOnlyTransaction(
-                session -> clientTemplate.findByEntityField(session, "name", "updatedName").stream()
-                        .map(Client::clone)
-                        .toList());
-
-        // then
-        assertThat(clientList).hasSize(1);
-        assertThat(clientList.getFirst()).usingRecursiveComparison().isEqualTo(savedClient);
-    }
-}
