--- L21-jpql/homework-template/src/test/java/ru/otus/crm/service/DbServiceClientTest.java
@@ -1,58 +1,0 @@
-package ru.otus.crm.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import ru.otus.base.AbstractHibernateTest;
-import ru.otus.crm.model.Client;
-
-@DisplayName("Демо работы с hibernate (с абстракциями) должно ")
-@SuppressWarnings("java:S125")
-class DbServiceClientTest extends AbstractHibernateTest {
-
-    @Test
-    @Disabled("Удалить при выполнении ДЗ")
-    @DisplayName(" корректно сохранять, изменять и загружать клиента")
-    void shouldCorrectSaveClient() {
-        // given
-        var client = new Client("Ivan");
-
-        // Это надо раскомментировать, у выполненного ДЗ, все тесты должны проходить
-        // Кроме удаления комментирования, тестовый класс менять нельзя
-        /*
-                var client = new Client(null, "Vasya", new Address(null, "AnyStreet"), List.of(new Phone(null, "13-555-22"),
-                        new Phone(null, "14-666-333")));
-        */
-
-        // when
-        var savedClient = dbServiceClient.saveClient(client);
-        System.out.println(savedClient);
-
-        // then
-        var loadedSavedClient = dbServiceClient.getClient(savedClient.getId());
-        assertThat(loadedSavedClient)
-                .isPresent()
-                .get()
-                .usingRecursiveComparison()
-                .isEqualTo(savedClient);
-
-        // when
-        var savedClientUpdated = loadedSavedClient.get();
-        savedClientUpdated.setName("updatedName");
-        dbServiceClient.saveClient(savedClientUpdated);
-
-        // then
-        var loadedClient = dbServiceClient.getClient(savedClientUpdated.getId());
-        assertThat(loadedClient).isPresent().get().usingRecursiveComparison().isEqualTo(savedClientUpdated);
-        System.out.println(loadedClient);
-
-        // when
-        var clientList = dbServiceClient.findAll();
-
-        // then
-        assertThat(clientList).hasSize(1);
-        assertThat(clientList.getFirst()).usingRecursiveComparison().isEqualTo(loadedClient.get());
-    }
-}
