+++ L21-jpql/homework/src/main/java/ru/otus/core/sessionmanager/DataBaseOperationException.java
