+++ L21-jpql/homework/src/main/java/ru/otus/core/sessionmanager/TransactionManager.java
