+++ L21-jpql/homework/src/main/java/ru/otus/crm/model/Address.java
@@ -1,0 +1,62 @@
+package ru.otus.crm.model;
+
+import jakarta.persistence.*;
+import java.util.Objects;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "address")
+public class Address implements Cloneable {
+
+    @Id
+    @SequenceGenerator(name = "address_gen", sequenceName = "address_seq", initialValue = 1, allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "address_gen")
+    private Long id;
+
+    @Column(name = "street")
+    private String street;
+
+    @OneToOne(fetch = FetchType.LAZY, mappedBy = "address")
+    private Client client;
+
+    public Address(String street) {
+        this.id = null;
+        this.street = street;
+    }
+
+    public Address(Long id, String street) {
+        this.id = id;
+        this.street = street;
+    }
+
+    @Override
+    @SuppressWarnings({"java:S2975", "java:S1182"})
+    public Address clone() {
+        return new Address(this.id, this.street);
+    }
+
+    @Override
+    public String toString() {
+        return "Address{" + "id=" + id + ", street='" + street + '\'' + '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Address address = (Address) o;
+        return Objects.equals(id, address.id) && Objects.equals(street, address.street);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, street);
+    }
+}
