+++ L21-jpql/homework/src/main/java/ru/otus/crm/model/Phone.java
@@ -1,0 +1,63 @@
+package ru.otus.crm.model;
+
+import jakarta.persistence.*;
+import java.util.Objects;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "phone")
+public class Phone implements Cloneable {
+
+    @Id
+    @SequenceGenerator(name = "phone_gen", sequenceName = "phone_seq", initialValue = 1, allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "phone_gen")
+    private Long id;
+
+    @Column(name = "number")
+    private String number;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "client_id")
+    private Client client;
+
+    public Phone(String number) {
+        this.id = null;
+        this.number = number;
+    }
+
+    public Phone(Long id, String number) {
+        this.id = id;
+        this.number = number;
+    }
+
+    @Override
+    @SuppressWarnings({"java:S2975", "java:S1182"})
+    public Phone clone() {
+        return new Phone(this.id, this.number);
+    }
+
+    @Override
+    public String toString() {
+        return "Phone{" + "id=" + id + ", number='" + number + '\'' + '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Phone phone = (Phone) o;
+        return Objects.equals(id, phone.id) && Objects.equals(number, phone.number);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, number);
+    }
+}
