+++ L21-jpql/homework/src/main/java/ru/otus/crm/service/DBServiceClient.java
