+++ L21-jpql/homework/src/main/java/ru/otus/crm/service/DbServiceClientImpl.java
