+++ settings.gradle.kts
@@ -32,7 +32,7 @@ include("L18-jdbc:homework")
 include("L19-rdbms")
 include("L20-hibernate")
 include("L21-jpql:class-demo")
-include("L21-jpql:homework-template")
+include("L21-jpql:homework")
 
 pluginManagement {
     val jgitver: String by settings
